create works (try/catch block)

+++ src/main/java/User.java
@@ -1,5 +1,5 @@
 public class User {
-    private long id = 0;     // should I set right here to 0????
+    private int id = 0;     // should I set right here to 0????
     private String userName;
     private String email;
     private String password;
@@ -10,11 +10,11 @@ public class User {
         this.password = password;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
